Added wildcard permission
Permissions are structured in groups: books.update, books.store….
For example for the index pages or for the menu links it is kind of
ugly to chain these as a user, who can only create new books, has to
see the book list to access this menu, as well as someone who can only
delete those (books is a bad example, transfer this to the user model).
The new AuthServiceProvider implementation allows to use
@can(‚people.*‘) or Gate::allows(‚people.*‘) to say: the user is
allowed to see or do this if he has any right of the people namespace
(e.g. people.update, people.store, people.delete)